refactor(editor): extract assets list from selected-component pane
